Apply MR suggestions to declare duration as int (same as node)
Respect the original int value from node->getDuration() (already
documented)

+++ include/megaapi_impl.h
@@ -3555,7 +3555,7 @@ protected:
     // File size
     m_off_t fileSize;
     // Media length in seconds (for media files)
-    size_t duration;
+    int duration;
 };
 
 class MegaTCPServer;
